Add files via upload
 Machine Learning Algorithms: Linear Regression

This repository contains an implementation of the Linear Regression algorithm as part of my machine learning studies. It includes  multiple linear regression model, along with code for data preprocessing, model training, evaluation, and visualization. The project serves as a foundational step in understanding and applying supervised learning techniques.

Feel free to explore, experiment, and contribute!